moved audit trail into a separate mixin
