+++ backend/users/views.py
@@ -1,16 +1,14 @@
 
 from django.shortcuts import get_object_or_404
-from django.utils import timezone
 from rest_framework import viewsets
 from rest_framework.response import Response
 
 from users.models import Userdata
 from users.serializers import UserdataSerializer
+from utils.mixins import AuditTrailMixin
 
-import logging
 
-
-class UserdataViewSet(viewsets.ModelViewSet):
+class UserdataViewSet(AuditTrailMixin, viewsets.ModelViewSet):
 	"""
 	API endpoint for CRUD actions over user data.
 
@@ -35,11 +33,3 @@ class UserdataViewSet(viewsets.ModelViewSet):
 
 	serializer_class = UserdataSerializer
 	queryset = Userdata.objects.all()
-
-	def _audit_trail(self, user=None, action=None, model='default', pk=None):
-		logger = logging.getLogger(f'audit.{model}.{action}')
-		logger.info(f'{timezone.now().isoformat()} {user} {action} {model} pk:{pk}')
-
-	def initial(self, request, *args, **kwargs):
-		self._audit_trail(user=request.user, action=self.action, model='userdata', pk=kwargs.get('pk'))
-		return super().initial(request, *args, **kwargs)
